console: a lapsed customer could leave over the API and not in the product
The hosted plan gate exempts five permissions so a customer whose subscription
has lapsed can still get out: billing.manage, data.export, organization.delete,
account.close and sessions.manage. That landed on the API in 7de5333a and its
own commit message said what was left: "these permissions are now reachable
over the API and still unreachable in the product". This is that half.

Every screen that reads them was gated. Settings is read from org.settings
under environments.view, which is precisely what the gate refuses, and
Shell.tsx sent anybody without hosted access to a standalone screen whose only
two actions were "Open plan and billing" and sign out. So the right existed in
the code and not in the product.

THE SHARP END was the confirmation. requestDeletion compares what you type
against the organization's SLUG (enterprise/deletion.ts:328) and every route
carrying the slug was gated, so a lapsed owner was asked to type a string the
product refused to show them. The slug is now on the screen, in its own field,
because it is the confirmation and not decoration.

SECOND CONFIRMATION, and the brief I was given had it wrong: account.close does
NOT confirm against the slug. It compares against c.actor.label
(enterprise.ts:279). Two different strings on one screen, and the dialogs are
asserted to name the right one each: "Type northwind-logistics to confirm" for
the deletion, "Type Priya Raghunathan to confirm" for the account.

Shell.tsx already had TWO lapsed branches and only one of them was a dead end.
Line 350 gave every path except /plan a standalone screen; line 369 gave /plan
a reduced shell with a real header and sign-out. The designed lapsed state
already existed and was scoped to one route, so this is an ALLOWLIST into it,
not a new screen: needsPlan && !LAPSED_PATHS.includes(pathname) for the
standalone, and the reduced shell for /plan plus /exits. One lapsed state
instead of two, and the exits inherit a header rather than being bolted onto a
dead end. The standalone now names the four exits and links them.

THE EXITS ARE MOVED, NOT COPIED. console/components/exits.tsx holds Sessions,
ExportCard, Deleting, CloseAccount and Fact, taken out of the settings page
rather than reimplemented beside it. A copy would drift, and the direction it
would drift is the one where the lapsed customer gets the older version of the
control they are entitled to. Deleting took `settings: Settings`, which is the
response of a route the lapsed customer is refused, so it now takes the four
facts it actually reads. `counts` is optional on that type: the read that
carries them is gated, and the confirmation says what is deleted without
numbers rather than saying "0 repositories", which would tell somebody about to
delete their organization that it is empty.

TWO DEFECTS FOUND WHILE BUILDING, both live on main and neither in scope:

/plan answers under billing.manage, which only an OWNER holds. So the old
lapsed screen offered a lapsed admin, member or viewer exactly one action, and
it was one that refuses them. The Plan link is now conditional and a non-owner
gets the exits as the primary action plus a sentence saying who can subscribe.

The one-time export link was thrown away one line after it was shown. The copy
beside it says "it is the only way to download the export once the organization
is gone, and it is shown once", which was true: deletion.request put the link
in component state and then called the parent's reload, which puts the parent
read back into its loading branch, unmounts the card, and destroys the link
before anybody could copy it. It renders the response now instead of
re-fetching. That one was CAUGHT BY THE BROWSER TEST, not by reading the diff,
and it is the assertion in that suite that has been watched to fail.

VERIFIED IN A REAL BROWSER, not from the code. Headless Chromium at 320, 390
and 1440, five states (owner, admin, member, viewer, a deletion in progress)
plus a healthy organization, with the page's own screenshots read back. No
horizontal scroll, no overflowing element, no console error, no animate-pulse.
The four exits were driven to completion and their requests inspected: 30 of 30
assertions, including that deletion.request carries the slug byte for byte and
account.close carries the label. The instrument was pointed at an injected
defect first and fired on all three checks, and the screenshot showed it, so
the clean result is from an instrument that has been watched to work.

The role table copy in console/lib/roles.ts now has a test, and it has been
watched to fail: removing account.close from viewer turns it red naming viewer.

STILL OPEN, and it is one string. The page makes one read, EXITS_READ in
app/(app)/exits/page.tsx, which must be the exempt read under account.close.
It is not on origin/ent-merge yet and enterprise-merge owns it. account.close
rather than deletion.status, because deletion.status is organization.delete and
an admin on a lapsed plan holds data.export and sessions.manage and not that,
so an owner-only read would have left an admin with no page. The name in the
file is a placeholder until enterprise-merge answers with the registered one.
